sonus-transcriber/requirements.txt
# Google Cloud dependencies
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-pubsub==2.28.0
psutil==7.0.0

# WhisperX dependencies
torch==2.6.0
torchaudio==2.6.0
faster-whisper==1.1.0
ctranslate2==4.4.0
transformers==4.49.0
pandas==2.2.3
setuptools>=65
nltk==3.9.1
whisperx==3.3.1
pyannote.audio==3.3.2
pytorch-lightning==2.5.0.post0
speechbrain==1.0.2
